refactor(editor): move form-field from inputs
