+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-array/form-field-array.component.spec.ts
@@ -8,7 +8,7 @@ describe('FormFieldArrayComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [FormFieldArrayComponent],
+      imports: [FormFieldArrayComponent],
     }).compileComponents()
 
     fixture = TestBed.createComponent(FormFieldArrayComponent)
