+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-file/form-field-file.component.html
