+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-file/form-field-file.component.spec.ts
@@ -1,6 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { FormFieldFileComponent } from './form-field-file.component'
+import { FormControl } from '@angular/forms'
 
 describe('FormFieldFileComponent', () => {
   let component: FormFieldFileComponent
@@ -8,11 +9,12 @@ describe('FormFieldFileComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [FormFieldFileComponent],
+      imports: [FormFieldFileComponent],
     }).compileComponents()
 
     fixture = TestBed.createComponent(FormFieldFileComponent)
     component = fixture.componentInstance
+    component.control = new FormControl()
     fixture.detectChanges()
   })
