+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-file/form-field-file.component.ts
@@ -1,11 +1,14 @@
+import { CommonModule } from '@angular/common'
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core'
-import { FormControl } from '@angular/forms'
+import { FormControl, ReactiveFormsModule } from '@angular/forms'
 
 @Component({
   selector: 'gn-ui-form-field-file',
   templateUrl: './form-field-file.component.html',
   styleUrls: ['./form-field-file.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [CommonModule, ReactiveFormsModule],
 })
 export class FormFieldFileComponent {
   @Input() control!: FormControl
