+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-object/form-field-object.component.spec.ts
@@ -8,7 +8,7 @@ describe('FormFieldObjectComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [FormFieldObjectComponent],
+      imports: [FormFieldObjectComponent],
     }).compileComponents()
 
     fixture = TestBed.createComponent(FormFieldObjectComponent)
