+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-object/form-field-object.component.ts
@@ -5,5 +5,6 @@ import { ChangeDetectionStrategy, Component } from '@angular/core'
   templateUrl: './form-field-object.component.html',
   styleUrls: ['./form-field-object.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
 })
 export class FormFieldObjectComponent {}
