+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-rich/form-field-rich.component.ts
@@ -1,11 +1,14 @@
+import { CommonModule } from '@angular/common'
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core'
-import { FormControl } from '@angular/forms'
+import { FormControl, ReactiveFormsModule } from '@angular/forms'
 
 @Component({
   selector: 'gn-ui-form-field-rich',
   templateUrl: './form-field-rich.component.html',
   styleUrls: ['./form-field-rich.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [CommonModule, ReactiveFormsModule],
 })
 export class FormFieldRichComponent {
   @Input() control!: FormControl
