+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-spatial-extent/form-field-spatial-extent.component.spec.ts
@@ -8,7 +8,7 @@ describe('FormFieldSpatialExtentComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [FormFieldSpatialExtentComponent],
+      imports: [FormFieldSpatialExtentComponent],
     }).compileComponents()
 
     fixture = TestBed.createComponent(FormFieldSpatialExtentComponent)
