+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field-temporal-extent/form-field-temporal-extent.component.spec.ts
@@ -8,7 +8,7 @@ describe('FormFieldTemporalExtentComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [FormFieldTemporalExtentComponent],
+      imports: [FormFieldTemporalExtentComponent],
     }).compileComponents()
 
     fixture = TestBed.createComponent(FormFieldTemporalExtentComponent)
