+++ libs/feature/editor/src/lib/components/record-form/form-field/form-field.component.ts
@@ -4,15 +4,39 @@ import {
   Input,
   Output,
 } from '@angular/core'
-import { FormFieldConfig } from './form-field.model'
-import { FormControl } from '@angular/forms'
+import { FormControl, ReactiveFormsModule } from '@angular/forms'
+import { MatIconModule } from '@angular/material/icon'
+import { TranslateModule } from '@ngx-translate/core'
 import { Observable } from 'rxjs'
+import { FormFieldArrayComponent } from './form-field-array/form-field-array.component'
+import { FormFieldFileComponent } from './form-field-file/form-field-file.component'
+import { FormFieldObjectComponent } from './form-field-object/form-field-object.component'
+import { FormFieldRichComponent } from './form-field-rich/form-field-rich.component'
+import { FormFieldSimpleComponent } from './form-field-simple/form-field-simple.component'
+import { FormFieldSpatialExtentComponent } from './form-field-spatial-extent/form-field-spatial-extent.component'
+import { FormFieldTemporalExtentComponent } from './form-field-temporal-extent/form-field-temporal-extent.component'
+import { FormFieldConfig } from './form-field.model'
+import { CommonModule } from '@angular/common'
 
 @Component({
   selector: 'gn-ui-form-field',
   templateUrl: './form-field.component.html',
   styleUrls: ['./form-field.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    FormFieldSimpleComponent,
+    FormFieldRichComponent,
+    FormFieldObjectComponent,
+    FormFieldSpatialExtentComponent,
+    FormFieldTemporalExtentComponent,
+    FormFieldFileComponent,
+    FormFieldArrayComponent,
+    CommonModule,
+    ReactiveFormsModule,
+    TranslateModule,
+    MatIconModule,
+  ],
 })
 export class FormFieldComponent {
   @Input() config: FormFieldConfig
