+++ libs/feature/editor/src/lib/feature-editor.module.ts
@@ -1,20 +1,20 @@
+import { CommonModule } from '@angular/common'
+import { HttpClientModule, HttpClientXsrfModule } from '@angular/common/http'
 import { NgModule } from '@angular/core'
 import { MatNativeDateModule } from '@angular/material/core'
 import { MatDatepickerModule } from '@angular/material/datepicker'
 import { MatFormFieldModule } from '@angular/material/form-field'
-import { WizardComponent } from './components/wizard/wizard.component'
-import { WizardFieldComponent } from './components/wizard-field/wizard-field.component'
 import { UiInputsModule } from '@geonetwork-ui/ui/inputs'
 import { UiWidgetsModule } from '@geonetwork-ui/ui/widgets'
-import { TranslateModule } from '@ngx-translate/core'
-import { WizardSummarizeComponent } from './components/wizard-summarize/wizard-summarize.component'
-import { CommonModule } from '@angular/common'
-import { HttpClientModule, HttpClientXsrfModule } from '@angular/common/http'
-import { StoreModule } from '@ngrx/store'
 import { EffectsModule } from '@ngrx/effects'
-import * as fromEditor from './+state/editor.reducer'
+import { StoreModule } from '@ngrx/store'
+import { TranslateModule } from '@ngx-translate/core'
 import { EditorEffects } from './+state/editor.effects'
 import { EditorFacade } from './+state/editor.facade'
+import * as fromEditor from './+state/editor.reducer'
+import { WizardFieldComponent } from './components/wizard-field/wizard-field.component'
+import { WizardSummarizeComponent } from './components/wizard-summarize/wizard-summarize.component'
+import { WizardComponent } from './components/wizard/wizard.component'
 
 @NgModule({
   declarations: [
@@ -42,5 +42,5 @@ import { EditorFacade } from './+state/editor.facade'
   providers: [EditorFacade],
 })
 export class FeatureEditorModule {}
-export * from './services/wizard.service'
 export * from './models/index'
+export * from './services/wizard.service'
