+++ libs/ui/inputs/src/index.ts
@@ -11,7 +11,6 @@ export * from './lib/button/button.component'
 export * from './lib/viewport-intersector/viewport-intersector.component'
 export * from './lib/check-toggle/check-toggle.component'
 export * from './lib/ui-inputs.module'
-export * from './lib/form-field'
 export * from './lib/copy-text-button/copy-text-button.component'
 export * from './lib/drag-and-drop-file-input/drag-and-drop-file-input.component'
 export * from './lib/navigation-button/navigation-button.component'
